% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myocardialInfarction.R
\docType{package}
\name{myocardialInfarction-package}
\alias{myocardialInfarction}
\alias{myocardialInfarction-package}
\title{myocardialInfarction: A Study Package for Generating Cohort Diagnostics}
\description{
A package that generates cohort diagnostics for cohort definitions included in the package.
}
\keyword{internal}
